Permisos para contadores en modulos de aviso obra y precios de etapas

+++ app/Http/Controllers/PrecioEtapaController.php
@@ -51,7 +51,7 @@ class PrecioEtapaController extends Controller
     //funcion para insertar en la tabla
     public function store(Request $request)
     {
-        if(!$request->ajax() || Auth::user()->rol_id == 11)return redirect('/');
+        if(!$request->ajax() || Auth::user()->rol_id == 11 || Auth::user()->rol_id == 9)return redirect('/');
         $precio_etapa = new Precio_etapa();
         $precio_etapa->fraccionamiento_id = $request->fraccionamiento_id;
         $precio_etapa->etapa_id = $request->etapa_id;
@@ -67,7 +67,7 @@ class PrecioEtapaController extends Controller
     //funcion para actualizar los datos
     public function update(Request $request)
     {
-        if(!$request->ajax() || Auth::user()->rol_id == 11)return redirect('/');
+        if(!$request->ajax() || Auth::user()->rol_id == 11 || Auth::user()->rol_id == 9)return redirect('/');
         //FindOrFail se utiliza para buscar lo que recibe de argumento
         $precio_etapa = Precio_etapa::findOrFail($request->id);
         $precio_etapa->fraccionamiento_id = $request->fraccionamiento_id;
@@ -98,7 +98,7 @@ class PrecioEtapaController extends Controller
 
     public function destroy(Request $request)
     {
-        if(!$request->ajax() || Auth::user()->rol_id == 11)return redirect('/');
+        if(!$request->ajax() || Auth::user()->rol_id == 11 || Auth::user()->rol_id == 9)return redirect('/');
         $precio_etapa = Precio_etapa::findOrFail($request->id);
         $precio_etapa->delete();
     }
